Ejercicio1.py

+++ Ejercicio1.py
@@ -2,6 +2,7 @@
 
 nombre = input("Cual es tu nombre?")
 Edad = int(input("Dame tu edad:"))
-Altura = int(input("Dame tu altura en metros: "))
+Altura = float(input("Dame tu altura en metros: "))
+
 
 print ("El usuario ",nombre, "de ", Edad, "y mide ", Altura, "metros.")
